Adding contact form with localStorge and also email footer like that

+++ js/form.js
@@ -21,9 +21,10 @@ document.body.addEventListener('click', (e) => {
 })
 
 ///?? form submit with User validation
-authForm.addEventListener('submit', function (e) {
+authForm?.addEventListener('submit', (e) => {
 
     e.preventDefault()
+    //* user information
     const user = {
         username: signIn ? undefined : username.value,
         email: email.value,
@@ -79,7 +80,18 @@ authForm.addEventListener('submit', function (e) {
             });
             return;
         }
-
+        // * confirmation passwrd
+        if (!signIn && confirmPassword.value.trim() !== password.value.trim()) {
+            // alert('Passwords do not match')
+            Swal.fire({
+                icon: "warning",
+                title: "Oops...",
+                text: `Passwords do not match`
+                
+            });
+    
+            return;
+        }
         users.push(user)
         localStorage.setItem('users', JSON.stringify(users));
         // alert('Registaration is successflly')
@@ -90,21 +102,12 @@ authForm.addEventListener('submit', function (e) {
             text: `You have successfully registered.`
 
         });
+        
 
         switchFormAuto()
     }
 
-    if (!signIn && confirmPassword.value.trim() !== password.value.trim()) {
-        // alert('Passwords do not match')
-        Swal.fire({
-            icon: "warning",
-            title: "Oops...",
-            text: `Passwords do not match`
-            
-        });
-
-        return;
-    }
+    
 
 })
 
@@ -142,3 +145,49 @@ function toggleMenu(){
     navbar.classList.toggle('active')
     loginButton.classList.toggle('active')
 }
+
+//* Email footer with locolStorage
+
+const footerEmail = document.querySelector('#footer-email')
+const footerBtn = document.querySelector('.telegram')
+// console.log(footerBtn)
+
+footerBtn.addEventListener('click', (e)=> {
+    e.preventDefault()
+    
+    let footerInput = footerEmail.value 
+    localStorage.setItem('footerEmail', footerInput)
+ 
+})
+
+// * Contact Page for Form
+
+const contactForm = document.querySelector('#contact-info')
+
+contactForm.addEventListener('submit', (e)=> {
+    e.preventDefault();
+    const firstName = document.getElementById('firstName')
+    const lastName = document.querySelector('#lastName')
+    const number = document.querySelector('#number')
+    const message = document.querySelector('#message');
+
+    let contactUser = {
+        firstName : firstName.value,
+        lastName : lastName.value,
+        email : email.value,
+        number : number.value,
+        message : message
+    }
+    let oldContact = JSON.parse(localStorage.getItem('ContactFormUser')) || []
+    console.log(oldContact)
+    let existContactUser = oldContact.find(CoUser => CoUser.email == email.value && CoUser.number == number.value)
+    console.log(existContactUser)
+    if(existContactUser){
+        alert(`This is already ${existContactUser.firstName} exist`)
+        return;
+    }
+    oldContact.push(contactUser)
+    localStorage.setItem('ContactFormUser', JSON.stringify(oldContact))
+
+  
+})
